@@ -3,37 +3,37 @@
   "version": "1.4.0",
   "private": true,
   "dependencies": {
-    "@cecalc/analisador-de-texto": "^4.0.5",
-    "@cecalc/utils": "^3.0.2",
-    "@fluentui/font-icons-mdl2": "^8.3.3",
-    "@fluentui/react": "^8.68.0",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^14.18.18",
-    "@types/office-js": "^1.0.253",
-    "@types/office-runtime": "^1.0.23",
-    "@types/react": "^17.0.45",
-    "@types/react-dom": "^17.0.17",
-    "@types/uuid": "^8.3.4",
+    "@cecalc/analisador-de-texto": "^4.2.0",
+    "@cecalc/utils": "^3.4.0",
+    "@fluentui/font-icons-mdl2": "^8.5.5",
+    "@fluentui/react": "^8.104.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.4",
+    "@types/node": "^18.11.17",
+    "@types/office-js": "^1.0.309",
+    "@types/office-runtime": "^1.0.29",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@types/uuid": "^9.0.0",
     "cross-env": "^7.0.3",
-    "es-check": "^6.2.1",
+    "es-check": "^7.0.1",
     "gh-pages": "^4.0.0",
     "lz-string": "^1.4.4",
     "npm-run-all": "^4.1.5",
-    "office-addin-cli": "^1.3.6",
-    "office-addin-dev-certs": "^1.7.8",
-    "office-addin-lint": "^2.0.1",
-    "office-addin-manifest": "^1.7.8",
-    "react": "^17.0.2",
+    "office-addin-cli": "^1.5.3",
+    "office-addin-dev-certs": "^1.11.0",
+    "office-addin-lint": "^2.2.3",
+    "office-addin-manifest": "^1.11.0",
+    "react": "^18.2.0",
     "react-app-polyfill": "^3.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-scripts": "^5.0.1",
-    "source-map-explorer": "^2.5.2",
-    "typescript": "^4.6.2",
-    "uuid": "^8.3.2",
-    "web-vitals": "^1.1.2"
+    "source-map-explorer": "^2.5.3",
+    "typescript": "^4.9.4",
+    "uuid": "^9.0.0",
+    "web-vitals": "^3.1.0"
   },
   "scripts": {
     "analyze": "source-map-explorer 'build/static/js/*.js'",
@@ -76,5 +76,8 @@
       "not dead",
       "not op_mini all"
     ]
+  },
+  "engines": {
+    "node": "~18.12.1"
   }
 }
